+++ views/components/common/menu.ejs
@@ -11,7 +11,7 @@
             <br>
             <li id="addNewContext" class="menu-item-divided"><a id="addcontextbutton" href="javascript:" class="whiten"> + add new graph&nbsp;&nbsp;</a><form action="/post" method="GET" class="pure-form" id="addcontextinput"><input type="text" id="addcontextname" size="8" class="pure-input"><input id="addcontextsubmit" type="submit" value="">&nbsp;&nbsp;</form></li>
             <li class="menu-item-divided"><a href="/<%= user.name %>/edit?showcontexts=1" class="whiten">see all graphs</a></li>
-            <% if (contextlist.length > 0) { %>
+            <% if (contextlist && contextlist.length > 0) { %>
             <% contextlist.forEach(function(contextlabel) { %>
                   <li class="menu-item-divided"><a href="/<%= user.name %>/<%= contextlabel[0] %>/edit"><%= contextlabel[0] %></a></li>
             <% }); %>
